build(quipu): bump pin to 0.3.23 (37bfc06a) — shacl gate ON, probes now pass in-process
Move the quipu git pin from 7f984b4e (0.2.0, 33 commits behind) to
37bfc06a (0.3.23, current main) and do the migration the old pin was
deliberately avoiding:

- tool_project now takes &mut Store: compute_code_ppr takes &mut, and the
  assembler's PPR site borrows self.quipu_store mutably (run_hybrid_search
  is already &mut self). tool_context stays &Store — unchanged.
- rusqlite 0.33 -> 0.40: libsqlite3-sys carries links="sqlite3", so bobbin
  and quipu must agree on one version. 0.40 dropped ToSql/FromSql for
  usize/u64; the COUNT(*) reads in storage/feedback.rs and the LIMIT bind
  in storage/sqlite/coupling.rs now go through i64 (lossless — counts are
  non-negative, LIMIT fits).
- lancedb 0.17 -> 0.27, arrow 53 -> 57: this is what makes quipu's shacl
  feature COMPILABLE at all. arrow-array 53.4.1 capped chrono <0.4.40 and
  every lancedb before 0.22.3 pinned chrono exactly, both irreconcilable
  with rudof_lib's chrono ^0.4.42; 0.27/arrow-57 is the pairing quipu
  itself builds against. API fallout: drop_table grew a namespace arg,
  add/create_table take impl Scannable (batch_to_reader now returns
  Box<dyn RecordBatchReader + Send>).
- quipu dep: features = ["shacl"], default-features = false. onnx stays
  OFF — cargo tree -i ort shows ort reached only from bobbin. This closes
  the missing write-gate (bobbin-di7): KNOWLEDGE_SHACL_ENABLED flips to
  true, and a SHACL refusal — which current quipu returns as Ok with
  conforms:false, not Err — is surfaced as an error on every write path
  (knowledge_knot, chunk push, coupling push, quarantine push) instead of
  masquerading as a landed write.
- Cargo.lock pins the whole rudof family to the versions in quipu's own
  lock (0.2.6/0.2.7/0.2.8): rudof_lib 0.2.9+ moved validate_shacl to a
  builder API this quipu rev does not compile against, and the family's
  loose cross-reqs break under mixed versions. Landmine documented in
  Cargo.toml.

What the bump unlocks, each proved by a new test against an in-process
store (the same crate the production path links, not a mock):

- replace_snapshot handled server-side: the chunks.rs probe passes and a
  re-push diffs instead of accumulating
  (snapshot_push_replaces_instead_of_accumulating).
- strict /knot graph routing (quipu 22b3569): the quarantine sentinel
  probe passes, inferred facts land in the registered crew:inferred plane
  and ROOT stays empty
  (push_inferred_lands_in_the_registered_plane_with_routing_enforced);
  an unregistered plane is refused naming camayoc's planes.py remedy.
- the SHACL gate actually refuses: a bobbin:Chunk without filePath is
  rejected with conforms:false
  (shacl_gate_is_compiled_in_and_refuses_a_violating_write) — the test
  that keeps shacl_validated:true honest.

The probes in chunks.rs/quarantine.rs STAY — they guard whatever store is
actually opened, not a particular pin; comments updated to say so.
Roadmap and quipu-integration plan updated: the pin row and quipu-080 are
closed, the chrono blocker text is superseded.

Gates: just check green, just test green (1109 tests, 0 failed).

+++ src/storage/sqlite/coupling.rs
@@ -20,7 +20,8 @@ impl MetadataStore {
                LIMIT ?2"#,
         )?;
         let results = stmt
-            .query_map(rusqlite::params![min_score, limit], |row| {
+            // rusqlite 0.40 dropped ToSql for usize; LIMIT fits i64.
+            .query_map(rusqlite::params![min_score, limit as i64], |row| {
                 Ok(FileCoupling {
                     file_a: row.get(0)?,
                     file_b: row.get(1)?,
